requirements.txt
mysql-connector-python==8.2.0
pymysql==1.1.0
pandas==2.1.4
sqlalchemy==2.0.23
pyspark==3.4.1
delta-spark==2.4.0
requests==2.31.0
boto3==1.34.34
